(TK-258) Add transform-config-rule-to-acl function
Without this patch there is no way to transform a rule expressed as
trapperkeeper configuration into an Access Control List data structure.
This is a problem because we need to feed ACL objects into the
authorization library.  This patch addresses the problem by implementing
the transformation function.  The implementation starts with an empty
acl, which itself is a sorted set, and then builds up the ACL by adding
named rules for each allow and deny entry in the TK configuration.