(TK-258) Implement wrap-handler ring middleware
Without this patch there the authorization service does not provide a
mechanism for dependants to wrap their own handlers with the
authorization check.  This patch addresses the problem by implementing
the `wrap-handler` service protocol function which wraps a ring handler
with middleware performing authorization.

A basic integration test provides the structure and intended use case
example of a trivial EchoReverseService wrapping its own handler with
the authorization check middleware.

Assertions and tests are expected to be filled out in future patches,
e.g. those that implement the specified rules format.