(TK-258) Refactor .toLowerCase into str/lower-case
Addresses https://github.com/puppetlabs/trapperkeeper-authorization/pull/2#discussion_r37570586

The data will be converted to lower-case even with the comment of
potentially not allowing lower-case at all because this will be a bit
easier for users to deal with.

Also, pass the configuration data value through the name function to
ensure it is a string representation of the information.  This is to
increase robustness dealing with str/lower-case