(TK-258) Re-use config->rule function
Without this patch the authorization service manually constructs the
data object for a rule and the associated access control list.  This is
a problem because the authorization library already includes
transformation functions for converting an individual rule expressed as
a map into a Rule data type suitable for use with the allowed? function.

This patch addresses the problem by copying the corresponding
config->rule function from the config namespace into the core namespace.
The tests and implementation are also changed to use the config->rule
function instead of the previous construction functions.