(TK-258) Add basic catalog regex authorization rule
Without this patch there are no integration tests that exercise the
behavior of the authorization library when there is a list of configured
rules.  This is a problem because there is only one trivial test that
exercises the default behavior of the library when there is an empty
list of configured rules.

This patch addresses the problem by introducing a single configured rule
that protects the catalog endpoint.  The intent is to allow only
requests which present an SSL client certificate matching the catalog
resource being accessed.

As written, this test passes because of the reverse DNS behavior in the
authorization library.  Because the request has no SSL certificate the
request should be denied.