(TK-258) Compute a request name solely from cert CN
Without this patch the authorization library falls back to using reverse
DNS to determine the name of the identity making the request.  This is a
problem because the security model is built around SSL.  This patch
addresses the problem by ensuring the name comes solely from the SSL
client certificate.

If there is no client certificate the name is set to the empty string.

The value of the [:authorization :authenticated?] request key will be
true only if the request has been verified as authentic, which is
typically done by validating the SSL client certificate.