(TK-258) Test default /puppet/v3/environments rule
Without this patch there is no test coverage of the first default rule
in the existing Ruby auth.conf [1]  This patch adds the coverage.  The
test currently specifies that the request should be allowed, which is
technically correct given that [TK-260][2] has not yet been implemented.

The structure is in place, however, for the `allowed?` method to check
the status of the `authentic?` sub-key.

This first rule should deny any unauthenticated request.  It does not,
however, which means the implementation needs to be updated.

[1]: https://github.com/puppetlabs/puppet/blob/826cee5/conf/auth.conf#L63-L65
[2]: https://tickets.puppetlabs.com/browse/TK-260