(TK-258) Deny unauthentic requests to /puppet/v3/environments
Without this patch requests which have no SSL certificate and as such
are not authentic are granted access to the /puppet/v3/environments
endpoint.  This is a problem because the default rules should prevent
access from unauthenticated clients.

This patch addresses the problem by changing the allowed? function to
check if the request has been authenticated or not.

The behavior of the function is to deny any requests which are not
authentic.  A future patch for TK-260 will augment this behavior to
allow unauthenticated requests.