(TK-258) Add authentic request integration tests
Without this patch there isn't coverage at the integration layer for the
use case of an authentic request with a valid SSL certificate.  This
patch addresses the problem by adding a test that covers this positive
use case in addition to the existing negative tests.

This patch refactors the integration tests to use the same
build-ring-handler approach that the ring-middleware tests employed.

The positive test is against the first default rule for the
`/puppet/v3/environments` URI endpoint.