(TK-258) Return HTTP 403 instead of 401
Authorization errors should be represented by HTTP 403, which is a
permanent error.  HTTP 401 is more for authentication, not
authorization, and a 401 is a temporary error instead of a permanent
error, the server is asking the client to try again.  These do not apply
to our authorization errors so 403 is more appropriate.