(TK-260) Add complete set default rules to integration tests
The default list of rules in the service integration test is intended to
model the default list of rules used by Puppet itself because this is a
real-world use case of the desired behavior of the library.  Without
this patch the default rules in the test contains only the first rule
from the old Ruby auth.conf [1] file.  This patch fills out a
translation of the rest of the rules from the master branch.

[1]: https://github.com/puppetlabs/puppet/blob/826cee5/conf/auth.conf