(TK-260) Make the request helper function variadic
Without this patch the request helper function will create a request
with the :ssl-client-cert key mapped to a value of nil.  This is a
problem because if this key is mapped then the value should be a
certificate according to the schema.  This patch addresses the problem
by not mapping the key if a certificate is not provided as an argument.

In order to accomplish this the patch moves the certificate argument to
the last position so that it is optional.