(TK-260) Add allow_unauthenticated rule attribute
Without this patch there is no mechanism to allow a request that has not
been authenticated.  This is a problem because some requests in Puppet,
e.g. the certificate status and CRL resources need to be accessible by
unauthenticated clients.

This patch addresses the problem by adding the allow_unauthenticated
attribute to an auth.conf rule.  If set to true, the request does not
need to have an identifying name (no name will be treated as the empty
string) and does not need to have a client certificate.