(TK-260) Error if allow_unauthenticated with allow/deny
Without this patch the system will not complain if there are allow/deny
statements inside of a rule with allow_unauthenticated true.  This is a
problem because we want to make sure this is treated as an error.  This
patch addresses the problem by adding logic to the core
`validate-auth-config-rule!` function.