(TK-268) Weight all deny ACEs before allow ACEs
Previously, the `ace-compare` function would sort ACEs in an ACL via
multiple criteria - by allow-all, then by exactness, etc.  This commit
changes the `ace-compare` function to weight all deny ACEs before
all allow ACEs, regardless of any other attributes of the pattern.

Since the ip portion of the match was removed, several of the associated
tests and functionality related to ip was removed as well.  This work
was slated to be done soon as part of TK-265 anyway so it seemed better
to go ahead and remove some of this code now rather than re-writing code
that would end up being thrown away soon after.