(TK-268) Removed `allow-all` references and updated docs
This commit removes references to `allow-all`.  Those had been in place
primarily to support the special ordering for "allow '*'" ACEs, which
was removed in an earlier commit in this branch.  "allow '*"" ACEs are
now treated more like other regex ACEs, similar to how "deny '*'" ACEs
are handled.

This commit also modifies the authorization-config.md to indicate that
requests are denied when both an allow and a deny ACE matches the name
for a matching rule.