(TK-261) Rule matching logging: produce a match report
Now upon calling `rules/allowed?` a match report is returned.
This datastructure describes all the match that happened, rule by
rule (describing rules that were matched, not-matched or skipped),
and the same for all ACL of a given rule.
This will allow to log with details exactly what happened during
matching and ACL check.