(PE-7742) Refactor schemas and query-param support
This commit merges the `AuthResultWithRequest` into the existing
`AuthorizationResult`, so now all authorization results contain the
request in question.

Query parameter destructuring has now been moved into the new
`authorization-check` function, and the existing
`wrap-with-authorization-check` has been refactored to call it. This
allows for `wrap-with-query-params` to be removed.

Added a test for the new `authorization-check` function which also
covers query parameter support and the newly returned (and updated)
request in the authorization result.