(maint) Check schema of just the certname
Previously the schema check for an ace-config entry would attempt to
validate that a map and not the value in the map was a string. As a map
is never a string, trying to use a certname entry in an allow/deny map
would not work. This commit corrects that to use the value and not the
map itself.