(maint) update dependency versions to modern versions
This updates clj-parent which brings in new versions of clj-rbac-client,
trapperkeeper, slingshot, prismatic/schema, kitchensink, ring-middleware,
ssl-urls and i18n.

Additionally ring-mock was moved into the dev-dependencies, and bouncycastle
was added as dev-dependencies for trapperkeeper-webserver-jetty9.

Clojure tools namespace was also updated to 1.4.1

At the time of the last update ring-codec was part of the ring-core
library and depended on java url decode for its implementation.  Since
then, it was moved into a separate ring-codec library and the implementation
changed to a more simple percent-encoding implementation. As a result,
there are no longer exceptions thrown by invalid encodings. A test that
was trying to detect that invalid decoding was changes to instead reflect
the new behavior.