(PE-36769) remove header warning from request->cert
This commit removes a warning log entry for the presense of a cert
in the request header when allow-header-cert-info is false.

This particular logging was deemed not terribly useful in general,
and the fix for agent auto-renew when CA proxy is present will
generate lots of log noise if it is present.